03_Chat_with_pdf_agent/requirements.txt
# dependencies for chat with pdf
python-dotenv
streamlit 
boto3
